requirements.txt
PyQt5>=5.15.0
mysql-connector-python>=8.0.0